third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/accessibility/autocomplete-attribute-expected.txt

Tests that autocompletions are computed correctly when editing the ARIA pane.


Running: testCheckedEmptyValue

Running: testCheckedFirstCharacter

Running: testRoleFirstCharacter
